+++ fedrec/communications_interfaces/zeroMQ_interfaces.py
@@ -1,7 +1,7 @@
 import zmq
 from zmq import Context
 from fedrec.utilities import registry
-from fedrec.communications import abstract_comm_manager as acm
+from fedrec.communications_interfaces import abstract_comm_manager as acm
 
 
 @registry.load("communications", "ZeroMQ")
